+++ irctest/controllers/unrealircd.py
@@ -3,6 +3,7 @@ import fcntl
 import functools
 from pathlib import Path
 import shutil
+import signal
 import subprocess
 import textwrap
 from typing import Callable, ContextManager, Iterator, Optional, Type
@@ -315,8 +316,9 @@ class UnrealircdController(BaseServerController, DirectoryBasedController):
         assert self.proc
 
         with _STARTSTOP_LOCK():
-            self.proc.kill()
-            self.proc.wait(5)  # wait for it to actually die
+            if not self._terminate_process_group(signal.SIGKILL):
+                self.proc.kill()
+            self.proc.wait(5)
             self.proc = None
